crystal-shards: add gitcache
Add gitcache to the crystal-shards container to increase the speed of
image building. add gitcache_host and gitcache_port in the
job.yaml file or use local gitcache service by default. gitcache_port defaults to port 5000.
gitcache_host and gitcache_port are the ip and port of the gitcache server.

Signed-off-by: cuiyili <cuiyili@huawei.com>