git_mirror: fix git-remote-http <defunct>
when use system() or %x() call shell command git, there would be a new
child process to do git command. And it seems the child process doesn't
clear everytime when it's finished. It needs to be waited by parent
process.

But it often comes out 'no child' error when wait in main thread, so I
trap SIGCHLD to SIG_IGN, and the init process will handle it.

Signed-off-by: Li Yuanchao <lyc163mail@163.com>