fix: scheduler: cut down log
[why]
when consume with "No job now", scheduler output 14 lines log.
when submit 1 job and consume it, scheduler output 78 lines log.
it is a bit of heavy for log collect and some logs are redundancy.

[how]
1.merge 4 lines to 1 line (form to json) for debug_message.
2.remove output "Done", it's same with kemal-frame meesage "200 GET/PUT..."
3.only 0 | 1 extra message for 1 API: if there has an active message, do
  not call debug_message

as default, there has 1 kemal-frame message for each API
  2020-10-20 09:40:28 UTC 200 GET /boot.ipxe/mac/0a-4c-34-18-0a-17 7.18ms
  <time in> <response status> <----request API---->  <response time span>

[now]
this patch cut 92 lines log to 26 line, interactive with 15 API.
examples: