sbin/es-jobs-mapping.sh: disable default dynamic mapping
[why]
By default, fields can be added dynamically to a document, or to
inner objects within a document, just by indexing a document
containing the new field.

However, there are many fields that we do not need, and "Limit of
total fields [1000] in index [jobs] has been exceeded" exception
will be reported when the number of fields exceeds 1000.

[how]
Disable default dynamic mapping.

Signed-off-by: Zhang Yuhang <zhangyuhang25@huawei.com>