add script for building my_info
for user deployed the whole compass-ci environment.
if they submit job in their own environment, the sched will still verify
the account with my_email, my_name and my_uuid.
however they cannot build the info in es and the default yaml file by
self.
to resolve this problem for these users, add this script for them to run
to build the info in es and default yaml file.

Signed-off-by: Luan Shengde <shdluan@163.com>