mail-robot: add check account exists for email
avoid re-assign accounts for a single email address

[why]
the mail-robot may assign a new account if user send 'apply account' ail
with the same email address. and this is a big waste of account resource.

[how]
when receive request for applying an account, check whether the email
already has already been assigned an account.
assign a new one if there is none
return the old one if there is already one and update info
mail robot will update infos if the mail has been set new:
  - my_name
  - my_commit_url
  - my_ssh_pubkey

Signed-off-by: Luan Shengde <shdluan@163.com>