sched: replace access_key to upload_dirs
We need to create the access_key in the dir which we want to upload
file, it is hard to handle that.
So we save the upload_dirs in redis, then we can set the field in
scheduler, the result-webdav container will read the value by job id and check the upload dir.

Signed-off-by: Cao Xueliang <caoxl78320@163.com>