mail-robot: optimize message message
optimize mail message for successfully applied account

[why]
its used linux command 'curl' to send request data when sending email.
so the ${HOME} and ${SHELL##*/} is treated as a system variable and
replaced with the variable value.
as a result, we will send a unexpected email.
to avoid this problem, need add '\\' before '$' in the old.

[how]
use rest-client instead of command 'curl' to send the request.

Signed-off-by: Luan Shengde <shdluan@163.com>