lib/job.cr: use an independent interface for submit
When we use Job.new, we need to do some extra processing only when we
submit the job.
Use interface "initialized?" to determine whether to submit a job.
However, there is a vulnerability.
If a job with all fields initialized is submitted,
the job can be successfully submitted without checking the account.

Signed-off-by: Wu Zhende <wuzhende666@163.com>