delimiter: bisect run script catch the exceptions and exit -1
Git bisect tag the commit good/bad/skip according to the bisect run
script exit code.

Exit code 0 is good, 125 is skip, the others between 1 and 127 is bad.

When the ruby script raise a exception exit code is 1,
git bisect will tag the commit to bad, but we should abort the bisect process.

So we catch the exceptions and exit -1 in bisect run script.

Signed-off-by: Cao Xueliang <caoxl78320@163.com>