container/submit: executable file submit
executable file for container: submit

file 'submit' is used as the command installed in the container.
it can be run as the real command submit itself in the container.
when you run it, it will calls the command 'submit' in the container.
add ln for it to PATH, you can run it at anywhere.

Signed-off-by: Luan Shengde <shdluan@163.com>