dracut-initrd: add support of custom bootstrap-$(arch).cgz
[Why]
When user want run job with their own custom bootstrap instead of
lkp-tests framework, they will generate a bootstrap-$(arch).cgz firstly,
and scheduler will give the bootstrap-$(arch).cgz path to testbox, then
testbox download the bootstrap-$(arch).cgz and use it as part of initrd.

So we add a step to handle bootstrap-$(arch).cgz, extract all of it to
$NEWROOT and boot from $NEWROOT.

[Addition]
The root dir in bootstrap-$(arch).cgz is /custom_bootstrap.

Signed-off-by: Yu Chuan <13186087857@163.com>