dracut-initrd: replace '-' to '_' of $os_version only for dracut stage
[Why]
In dracut stage, we use $os_version to create logical volume, and copy
our rootfs to the logical volume.

And if there have char '-' between the logical volume name, it will be
transferred to '--' when system create logical volume.

So we replace '-' to '_' of $os_version only for dracut stage.

Signed-off-by: Yu Chuan <13186087857@163.com>