lib/config_account.rb: fix read my_info failed bugs
this tool is used to read account info from es and generate account yaml
files.

but as soon as the ES is started, the ES may still didn't read the
database from the storage, and it may read the account info in failure,
and so it will be failed to generate the right account yaml files.

add check for my_info for some times to read account info from the ES.
case read none until the loop times over, the check will break.

Signed-off-by: Luan Shengde <shdluan@163.com>