os_mount=local: add support os_lv field
[Why]
Currently, when job want use os_mount=local, if job need reboot and use
the rootfs data before reboot, we need do the following steps:
1. submit 1.job with the two key fields:
   - os_mount: local
   - boot_lv_suffix: yuchuan_iperf
2. reboot/release the testbox of 1.job.
3. submit 2.job with the tow key fields:
   - os_mount: local
   - src_lv_suffix: yuchuan_iperf
Then, 2.job will use the rootfs data of 1.job.

It is true that using pairing can serve as a version log of rootfs data,
but it can be a bit cumbersome when a job only needs a copy of rootfs
data to use all the time.

[How]
So we support a new field: os_lv.
After add that, the above flow will be the follow:
1. submit 1.job with the two key fields:
   - os_mount: local
   - os_lv: yuchuan_iperf
2. reboot/release the testbox of 1.job.
3. submit 2.job with the tow key fields:
   - os_mount: local
   - os_lv: yuchuan_iperf
Then, 2.job will use the rootfs data of 1.job.

Signed-off-by: Yu Chuan <13186087857@163.com>