docker/docker.rb: fix untimely restart
[error]
"msg":"Error response from daemon: removal of container z9.5070420 is ready in progress"

[why]
When the preceding error occurs,
docker will be stuck at the "docker rm -f xxx" step for a long time.
As a result, the message following the MQ can not be consumed in time,
and the docker can not be restarted in time.

Signed-off-by: Wu Zhende <wuzhende666@163.com>