lib/lifecycle: change physical machine reboot queue
dc and vm are deployed on different physical machines.
The restart service is a thread of each multi-qemu or multi-docker on each physical machine.
Therefore, the restart queue must be specific to the thread of the physical machine.
However, physical machines are restarted on the IBMC service.
The restart queue should be set to one.

Signed-off-by: Wu Zhende <wuzhende666@163.com>