kvm.sh: optimize the logic of disk preparation
[Why]
in lkp_src/setup/md, when create raid, it will add --assume-clean param,
this param will keep the old data of disk.

but this will caused the usage of disk will be greater and greater, so
we need add the following logic:
- clean the hdd_partition for every job;
- keep the rootfs_disk;

and further, the reclaim of rootfs_disk will be fulfilled in later
patch.

Signed-off-by: Yu Chuan <13186087857@163.com>