scheduler/elasticsearch_client.cr: add a new function update_job
The old function to update job content is "set_job_content".
But it will get the job first.
There are places where I don't want to do a query first,
because I've already done it.
And it's wasteful to do it again.
So I add this new function only do update.

Signed-off-by: Wu Zhende <wuzhende666@163.com>