monitoring/amqp.cr: fix exclusive the fiber
There are three fibers in monitoring service:
  - spawn deal serial-logging
  - spawn deal docker-logging
  - spawn kemal run
Three fibers will compete with CPU to do its own thing.
If kemal obtains the execution permission, it will be released soon.
However, if the other two get execute permission,
they will consume all the messages in the current MQ queue before releasing the permission.
If serial-logging get CPU, and there are 10000000 messages in rabbitmq queue,
it will hog the CPU for a long time, causing the other two fibers to wait.

Signed-off-by: Wu Zhende <wuzhende666@163.com>