lib/parse_serial_logs.cr: improve processing efficiency
[Why]
As the cluster size increases, the number of serial port logs becomes larger.
And logs are now being collected faster than the service can process them.
This can cause logs to accumulate in the RabbitMQ queue
and eventually cause the RabbitMQ service to collapse.

[Analyse]
I added time to print each step in the function that handles the serial log.
To see where the performance bottlenecks are.
output(unit: us):
{
  "delete_host_start" => 55.96156100910783,
  "check_save"        => 5.373575258654693,
  "delete_host_end"   => 28.90111722763539,
  "match_job_id"      => 3.1805337069773025,
  "find_job"          => 0.7753996618464105,
  "dump_cache"        => 1.0356967130094934,
  "cache_dmesg"       => 3.178479275962366
}
According to the above results,
we can see that the bottleneck is mainly in two delete_hosts,
so we put the optimization direction on delete_host.

[Basis]
After testing,
"message.includes?(pattern)" is about ten times faster than
"message.match(/.*(?<signal>#{pattern})/)"

[Result]
This is the result of using optimized code:
{
  "delete_host_start" => 5.6527033634739485,
  "check_save"        => 10.23306258904446,
  "delete_host_end"   => 3.9768935212232766,
  "match_job_id"      => 5.6382087210679,
  "find_job"          => 0.7921993629628897,
  "dump_cache"        => 0.9519881674566717,
  "cache_dmesg"       => 4.635072723158488
}

Signed-off-by: Wu Zhende <wuzhende666@163.com>