qemu/kvm.sh: fix dmesg chaos
The dmesg log of one job is mixed with the logs of other jobs.
Service get log like this:
  {"message": "", "serial_path": "/srv/cci/serial/logs/vm-2p8g.taishan...", "time": ""}
The service does not know which job this log belongs to.
Until this message is received:
  "initrd http://172.168.131.113:3000/job_initrd_tmpfs/crystal.3000726/job.cgz"
Therefore, this line cannot appear in the serial port log of the VM and
the physical machine at the same time.
Otherwise, the log of both machines will be written to the dmesg of the same job

Signed-off-by: Wu Zhende <wuzhende666@163.com>