service/scheduler: support depend rpmbuild
[Background]
The existing functionality is to trigger rpmbuild tests for a package when the source code of the package is updated.
Only the build problems of this package can be found, but there are various dependencies between packages.
A package update may introduce compatibility issues that cause other packages that depend on that package to fail to build.
So what we do is when a package's source code is updated, trigger that package and its dependencies to run rpmbuild tests.

Single package build process:
- git-mirror
- auto-submit
- run job（rpmbuild）
- rpm-repo（update repo）

Depend on the package build process:
A depends on B
B rpmbuild:
- After package B is successfully constructed, upload the RPM package to the server
- Find reverse dependencies: dnf repoquery --whatrequires B
- Notifies the scheduler that the RPM package has been uploaded(curl http://$SCHED_HOST:$SCHED_PORT/rpmbuild/depend)
  POST json:
  {
    "depend_job_id": xx,
    "depend_rpm_dest": xx,
    "reverse_depends": "A C D"
  }

Scheduler:
- api /rpmbuild/depend:
  - Check whether RPM packages need to be moved based on upstream_branch (master only needs to be moved)
  - Assemble job information
  - submit job
- api /download:
  Can download RPM packages

A rpmbuild:
- get B rpm packages：curl http://$SCHED_HOST:$SCHED_PORT/download?path=$depend_rpm_dest -O
- yum localinstall -y  *.rpm

Signed-off-by: Wu Zhende <wuzhende666@163.com>