scheduler: fix the key serial log missing for the x86_64 testbox
[Why]
When multiple consoles are listed output is sent to all consoles and
input is taken from the last listed console. The last console is the one
Linux uses as the /dev/console device.

The output of the init process appears only on the last console.

So we need exchange the order of multiple kernel console params.

[Refer]
https://tldp.org/HOWTO/Remote-Serial-Console-HOWTO/configure-kernel.html#CONFIGURE-KERNEL-SYNTAX

Signed-off-by: Yu Chuan <13186087857@163.com>