providers: qemu: put the chdir step in qemu.sh
[Why]
We use multi threads to start a vm, and we need chdir to vm's root dir.

If we change dir in one thread, then the pwd(current dir) of main thread
and other threads will all be changed.

So we put the chdir step to qemu.sh, and because of qemu.sh will be
executed in 'system' cmd, and 'system' cmd will execute the str in a
subshell. Then the pwd(current dir) of main thread and other threads
will not be changed.

Signed-off-by: Yu Chuan <13186087857@163.com>