providers: fix establish repeated connections to RabbitMQ
Bunny client automatically reconnects to RabbitMQ when it is abnormally disconnected.
The original code logic catches connection failures and
reestablishes a new connection every 5 seconds.
This will result in multiple connections being re-established once the
RabbitMQ server is available again.
This will overload rabbitMQ with connections and cause the load to crash.

Signed-off-by: Wu Zhende <wuzhende666@163.com>