service/monitoring: split into two independent services
[Why]
The old monitoring service has two functions.
One is to monitor docker logs and the other is to
write serial port logs to the dmesg file of the corresponding job.
Because serial port logs contain a large amount of data,
problems such as log accumulation may occur.
Thus, the function of "submit -m -c" is affected
and the borrowed machine cannot be logged in in time.
So split these two functions into two separate services that do
not interfere with each other.
The advantage of this is to ensure the stability of
the function of "submit -m -c" and to be more focused when encountering problems.

[Now]
monitoring-20001: monitor docker logs, support for "submit -m".
serial-logging: generate file dmesg, monitor machine crash.

Signed-off-by: Wu Zhende <wuzhende666@163.com>