mail-robot: add new key for user info
sometimes we need to know if a user has the way to access account-vm
server, add new key account_vm for user info.

if the key's value is true, it means the user has the way to login the
account-vm server.
if the key's value is false, it means user has no sense with the
account-vm server.

Signed-off-by: Luan Shengde <shdluan@163.com>