container: gem install elasticsearch:7.11.1
For user locally deploy compass-ci, error message as follows:
root@taishan200-2280-2s64p-128g--a1002 /c/compass-ci/container/scheduler# ./start
The client is unable to verify that the server is Elasticsearch. Some functionality may not be compatible if the server is running an unsupported product.
Traceback (most recent call last):
        9: from /c/compass-ci/sbin/config_account_yaml.rb:8:in `<main>'
        8: from /c/compass-ci/lib/config_account.rb:17:in `config_yaml'
        7: from /c/compass-ci/lib/config_account.rb:17:in `new'
        6: from /c/compass-ci/lib/build_my_info_client.rb:26:in `initialize'
        5: from /c/compass-ci/lib/build_my_info_client.rb:26:in `new'
        4: from /c/compass-ci/lib/es_query.rb:14:in `initialize'
        3: from /usr/lib/ruby/gems/2.7.0/gems/elasticsearch-api-8.0.0/lib/elasticsearch/api/actions/ping.rb:39:in `ping'
        2: from /usr/lib/ruby/gems/2.7.0/gems/elasticsearch-8.0.0/lib/elasticsearch.rb:70:in `method_missing'
        1: from /usr/lib/ruby/gems/2.7.0/gems/elasticsearch-8.0.0/lib/elasticsearch.rb:104:in `verify_elasticsearch'
/usr/lib/ruby/gems/2.7.0/gems/elasticsearch-8.0.0/lib/elasticsearch.rb:115:in `verify_with_version_or_header':
The client noticed that the server is not Elasticsearch and we do not support this unknown product. (Elasticsearch::UnsupportedProductError)
The client is unable to verify that the server is Elasticsearch. Some functionality may not be compatible if the server is running an unsupported product.

Because $CCI_SRC/container/es use elasticsearch:7.11.1, but gem install elasticsearch:8.0.0 if not specify the version,
the error is caused by inconsistent versions.

Signed-off-by: Liu Yinsi <liuyinsi@huawei.com>