container/repo-mirror: rename start to run
Repo-mirror is a tunasync service, it is used to build a local mirror station.
It is very convenient to install various software dependencies.

refer to:
https://github.com/tuna/tunasync
https://github.com/tuna/tunasync/blob/master/docs/zh_CN/get_started.md

For user locally deploy compass-ci, because of many mirrors in workers.conf,
it will take up huge amount of storage space(/srv/os-repo), as follows:

/srv/os-repo% du -sh * 2>&/dev/null
65G     archlinux
80G     archlinuxarm
318G    centos
190G    centos-altarch
1.8T    debian
210G    debian-security
271G    epel
11G     fedora
16K     lost+found
164G    openEuler-20.03-LTS-SP1
3.1G    opensuse
1.7T    ubuntu
2.1T    ubuntu-ports

When rename start to run, will not start this container during deploy compass-ci,
to avoid this issue.

Signed-off-by: Liu Yinsi <liuyinsi@huawei.com>