sparrow/0-package/common: download built images
For user locally deploy compass-ci, the network of some users is
unstable, the image build is slow or even fails.

So we provide built images for users.
https://repo.oepkgs.net/openEuler/compass-ci/cci-deps/docker/$(arch)/images.cgz,

But we don't provide es and logging-es images, because its password is
randomly generated in user's server(/etc/compass-ci/passwd.yaml), the
password of the built images are different from the local password of
the user, so user must build es and logging-es images by themselves.

Signed-off-by: Liu Yinsi <liuyinsi@huawei.com>