container/osimage: refactor this module code that generates initramfs
[why]
There are a lot of duplicate codes in this module code, I have
subtracted this duplicate codes, but it doesn't affect the original functionality.

And adding a unified entry docker-build-initrd for this module.

[usage]
./docker-build-initrd --os_name openeuler os_version 22.03-LTS
Generate initrd of the specified docker_image.

Mandatory arguments to long options are mandatory for short options too.
  -n, --name            the name of generated initramfs.
      --arch            the arch of docker_image.
      --os_version      the version of openeuler.
      --root_passwd     specify the root password of initrd, the default password is "test".
  -h, --help            display this help and exit
  -d, --debug           show the DEBUG level log, default is INFO level.

Users can run docker-build-initrd with parameters to generate their
own initrd

Signed-off-by: Li Ping <1477412247@qq.com>